pyproject.toml
[project]
name = "model-diversity"
version = "0.1.0"
description = "Is model diversity worth it?"
readme = "README.md"
authors = [
    { name = "David Koleczek", email = "45405824+DavidKoleczek@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.52",
    "codex-app-server-sdk>=0.3.2",
    "huggingface-hub>=1.3",
    "interop-router==0.1.12",
    "pyarrow>=19.0",
    "python-liquid>=2.1.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["uv_build>=0.9.15,<1.0"]
build-backend = "uv_build"

[tool.uv]
exclude-newer = "5 days"

[dependency-groups]
dev = [
    "ipykernel>=7.2",
    "pre-commit>=4.2",
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "ruff>=0.15",
    "ty>=0.0.21",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "F",    # pyflakes
    "E",    # pycodestyle
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "RUF",  # ruff
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "ISC",  # flake8-implicit-str-concat
    "PIE",  # flake8-pie
    "PT",   # flake-pytest-style
    "PTH",  # flake8-use-pathlib
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
]
ignore = [
    "E501",   # line-too-long (formatter makes best-effort attempt)
    # The following rules conflict with the formatter, see: https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",   # tab-indentation
    "E111",   # indentation-with-invalid-multiple
    "E114",   # indentation-with-invalid-multiple-comment
    "E117",   # over-indented
    "ISC002", # multi-line-implicit-string-concatenation
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
